Move sound effects to  as CONSTANTS
